feat: extract joystick, camera, and paint hooks
