+++ src/components/MemoryPalace.jsx
@@ -1,8 +1,8 @@
 import React, { useRef, useState, useEffect, forwardRef, useImperativeHandle } from 'react'
 import * as THREE from 'three'
-import nipplejs from 'nipplejs'
 import SimpleParticleManager from '../utils/SimpleParticleManager'
 import SettingsManager from '../services/SettingsManager'
+import { useJoystick, usePaintMode, useCameraControls } from '../hooks'
 
 const MemoryPalace = forwardRef(({ 
   wireframeEnabled = false, 
@@ -25,8 +25,6 @@ const MemoryPalace = forwardRef(({
   const rendererRef = useRef(null)
   const cameraRef = useRef(null)
   const wireframeSphereRef = useRef(null)
-  const nippleManagerRef = useRef(null)
-  const nippleContainerRef = useRef(null)
   const skyboxSphereRef = useRef(null)
   const skyboxMaterialRef = useRef(null)
   const particleManagerRef = useRef(null)
@@ -34,1352 +32,41 @@ const MemoryPalace = forwardRef(({
   const offScreenIndicatorsRef = useRef(new Map())
   const animationFrameRef = useRef(null)
   const settingsManagerRef = useRef(new SettingsManager())
-  
-  // Paint mode refs
-  const paintCanvasRef = useRef(null)
-  const paintTextureRef = useRef(null)
-  const paintContextRef = useRef(null)
-  const [paintedGroups, setPaintedGroups] = useState(new Map()) // Store painted object groups
-  const paintModeEnabledRef = useRef(paintModeEnabled) // Ref for current paint mode state
-  const paintInitializedRef = useRef(false) // Prevent double initialization in strict mode
-  
-  // Camera rotation state - needs to be accessible across all functions
-  const cameraRotationRef = useRef({ yaw: 0, pitch: 0 })
 
-  // Paint mode functions
-  const initializePaintCanvas = () => {
-    if (paintCanvasRef.current || paintInitializedRef.current) return // Already initialized or initializing
-    
-    paintInitializedRef.current = true // Prevent double initialization
-    console.log('[MemoryPalace] Initializing paint canvas system')
-    
-    // Create canvas for painting - high resolution for detail
-    const canvas = document.createElement('canvas')
-    canvas.width = 4096 // High resolution equirectangular
-    canvas.height = 2048
-    
-    const context = canvas.getContext('2d')
-    
-    // Initialize with subtle visible base pattern so we can see if the texture is rendering
-    // Start with very subtle transparent overlay
-    context.fillStyle = 'rgba(0, 0, 0, 0)' // Fully transparent base
-    context.fillRect(0, 0, canvas.width, canvas.height)
-    
-    // Add very subtle grid pattern to confirm texture visibility 
-    context.strokeStyle = 'rgba(255, 255, 255, 0.05)' // Very faint white lines
-    context.lineWidth = 2
-    
-    // Add sparse grid lines for debugging visibility
-    const gridSpacing = 200
-    for (let x = 0; x < canvas.width; x += gridSpacing) {
-      context.beginPath()
-      context.moveTo(x, 0)
-      context.lineTo(x, canvas.height)
-      context.stroke()
-    }
-    for (let y = 0; y < canvas.height; y += gridSpacing) {
-      context.beginPath()
-      context.moveTo(0, y)
-      context.lineTo(canvas.width, y)
-      context.stroke()
-    }
-    
-    // Store references
-    paintCanvasRef.current = canvas
-    paintContextRef.current = context
-    
-    // Create Three.js texture from canvas
-    const paintTexture = new THREE.CanvasTexture(canvas)
-    paintTexture.mapping = THREE.EquirectangularReflectionMapping
-    paintTexture.wrapS = THREE.RepeatWrapping
-    paintTexture.wrapT = THREE.ClampToEdgeWrapping
-    paintTexture.offset.x = 0.5 // Apply same 180° offset as all skybox textures for coordinate system alignment
-    paintTexture.needsUpdate = true
-    
-    paintTextureRef.current = paintTexture
-    
-    console.log('[MemoryPalace] Paint canvas system initialized with subtle grid pattern for visibility testing')
-  }
-
-  const enablePaintMode = () => {
-    if (!skyboxMaterialRef.current || !skyboxSphereRef.current) return
-    
-    console.log('[MemoryPalace] Enabling paint mode')
-    
-    // Initialize paint canvas if not already done
-    initializePaintCanvas()
-    
-    // Create a separate paint sphere that overlays on the skybox
-    if (paintTextureRef.current && !skyboxSphereRef.current.userData.paintSphere) {
-      // Store original skybox state
-      if (!skyboxMaterialRef.current.userData.originalMap) {
-        skyboxMaterialRef.current.userData.originalMap = skyboxMaterialRef.current.map
-        skyboxMaterialRef.current.userData.originalTransparent = skyboxMaterialRef.current.transparent
-      }
-      
-      console.log('[MemoryPalace] Paint mode enabled - using actual paint canvas texture')
-      
-      // Create paint material with actual paint canvas texture
-      const paintMaterial = new THREE.MeshBasicMaterial({
-        map: paintTextureRef.current, // Use actual paint texture for painting
-        transparent: true,
-        opacity: 0.6,
-        depthTest: false,
-        depthWrite: false
-      })
-      
-      // Create paint sphere inside the skybox for inside viewing
-      const paintGeometry = new THREE.SphereGeometry(499, 60, 40) // Inside skybox (500)
-      paintGeometry.scale(-1, 1, 1)
-      
-      const paintSphere = new THREE.Mesh(paintGeometry, paintMaterial)
-      sceneRef.current.add(paintSphere)
-      
-      // Store reference for cleanup
-      skyboxSphereRef.current.userData.paintSphere = paintSphere
-      skyboxSphereRef.current.userData.paintMaterial = paintMaterial
-      skyboxSphereRef.current.userData.paintGeometry = paintGeometry
-      
-      console.log('[MemoryPalace] Paint overlay sphere created with actual paint canvas texture - should show subtle grid pattern')
-    }
-  }
-
-  const disablePaintMode = () => {
-    if (!skyboxMaterialRef.current || !skyboxSphereRef.current) return
-    
-    console.log('[MemoryPalace] Disabling paint mode')
-    
-    // In creation mode, make painted areas available to LLM instead of directly creating objects
-    if (creationModeActive) {
-      console.log('[MemoryPalace] Creation mode active - making painted areas available to LLM')
-      exposePaintedAreasToLLM()
-    } else {
-      // Normal paint mode - process painted areas into objects directly
-      processPaintedAreasIntoObjects()
-    }
-    
-    // Remove paint sphere overlay
-    if (skyboxSphereRef.current.userData.paintSphere) {
-      const paintSphere = skyboxSphereRef.current.userData.paintSphere
-      const paintMaterial = skyboxSphereRef.current.userData.paintMaterial
-      const paintGeometry = skyboxSphereRef.current.userData.paintGeometry
-      
-      // Remove from scene
-      sceneRef.current.remove(paintSphere)
-      
-      // Dispose resources
-      if (paintGeometry) paintGeometry.dispose()
-      if (paintMaterial) paintMaterial.dispose()
-      // Note: Don't dispose paintTextureRef.current as it's managed separately
-      
-      // Clear references
-      skyboxSphereRef.current.userData.paintSphere = null
-      skyboxSphereRef.current.userData.paintMaterial = null
-      skyboxSphereRef.current.userData.paintGeometry = null
-      
-      console.log('[MemoryPalace] Paint overlay sphere removed')
-    }
-  }
-
-  // Cleanup function to reduce memory usage by removing detailed stroke data from stored objects
-  const cleanupStrokeData = (objects = []) => {
-    console.log('[MemoryPalace] Cleaning up stroke data from objects to reduce memory usage')
-    let cleanupCount = 0
-    
-    objects.forEach(obj => {
-      if (obj.paintData && obj.paintData.areas && Array.isArray(obj.paintData.areas)) {
-        // Replace detailed stroke data with summary statistics
-        const areas = obj.paintData.areas
-        const summary = {
-          strokeCount: areas.length,
-          totalArea: areas.reduce((sum, area) => sum + (area.size * area.size || 0), 0),
-          averageStrokeSize: areas.length > 0 ? areas.reduce((sum, area) => sum + (area.size || 0), 0) / areas.length : 0
-        }
-        
-        // Remove detailed stroke data and replace with summary
-        delete obj.paintData.areas
-        obj.paintData.strokeSummary = summary
-        obj.paintData.cleaned = true
-        cleanupCount++
-        
-        console.log(`[MemoryPalace] Cleaned stroke data for object ${obj.id || obj.name}: ${areas.length} strokes → summary`)
-      }
-    })
-    
-    console.log(`[MemoryPalace] Stroke cleanup complete: ${cleanupCount} objects processed`)
-    return cleanupCount
-  }
-
-  const exposePaintedAreasToLLM = () => {
-    console.log('[MemoryPalace] Exposing painted areas to LLM for creation mode')
-    
-    if (paintedGroups.size === 0) {
-      console.log('[MemoryPalace] No painted areas to expose')
-      return
-    }
-    
-    // Group painted areas by proximity to create contiguous objects/doors
-    const paintedAreas = Array.from(paintedGroups.values())
-    const groups = groupContiguousPaintAreas(paintedAreas)
-    
-    console.log(`[MemoryPalace] Exposing ${groups.length} painted area groups to LLM`)
-    
-    // Convert groups to LLM-friendly format
-    const paintedAreaData = groups.map((group, index) => {
-      const groupData = createObjectFromPaintGroup(group, index + 1, paintModeType)
-      
-      return {
-        id: `painted_area_${index + 1}`,
-        name: groupData.name,
-        information: groupData.information,
-        position: groupData.position,
-        dimensions: groupData.dimensions,
-        paintedType: paintModeType, // 'objects' or 'doors'
-        strokeCount: group.length,
-        canvasCenter: groupData.canvasCenter,
-        worldPosition: groupData.position
-      }
-    })
-    
-    // Pass painted area data through callback instead of global
-    if (onPaintedAreasChange) {
-      onPaintedAreasChange(paintedAreaData)
-    }
-    
-    console.log('[MemoryPalace] Painted area data passed through callback:', paintedAreaData)
-    
-    // Don't clear painted groups yet - let LLM process them first
-    // setPaintedGroups(new Map())
-  }
-
-  const processPaintedAreasIntoObjects = () => {
-    console.log('[MemoryPalace] Processing painted areas into objects/doors, mode:', paintModeType)
-    
-    if (!paintCanvasRef.current || paintedGroups.size === 0) {
-      console.log('[MemoryPalace] No painted areas to process')
-      return
-    }
-    
-    // Group painted areas by proximity to create contiguous objects/doors
-    const paintedAreas = Array.from(paintedGroups.values())
-    const groups = groupContiguousPaintAreas(paintedAreas)
-    
-    console.log(`[MemoryPalace] Found ${groups.length} contiguous paint groups from ${paintedAreas.length} paint areas`)
-    
-    // Convert each group to a memory palace object or door based on paint mode type
-    groups.forEach((group, index) => {
-      const groupData = createObjectFromPaintGroup(group, index + 1, paintModeType)
-      
-      // Use AI properties if available, otherwise use defaults
-      const finalName = aiObjectProperties?.name || groupData.name
-      const finalInformation = aiObjectProperties?.information || aiObjectProperties?.description || groupData.information
-      const finalType = aiObjectProperties?.type || (paintModeType === 'doors' ? 'door' : 'object')
-      
-      if (finalType === 'door' || paintModeType === 'doors') {
-        // Create painted door with proper type structure
-        const paintedDoorParams = {
-          name: finalName,
-          type: 'door',
-          description: finalInformation,
-          information: finalInformation,
-          position: groupData.position,
-          targetRoomId: aiObjectProperties?.targetRoomId || '', // Will need to be configured later
-          isPaintedDoor: true,
-          paintData: {
-            // Store only essential data after object creation - remove detailed stroke data
-            dimensions: groupData.dimensions,
-            canvasPosition: groupData.canvasCenter,
-            color: 'rgba(0, 0, 255, 0.9)', // Blue for doors
-            aiGenerated: !!aiObjectProperties,
-            // Note: areas with full stroke data removed to reduce serialization overhead
-            strokeCount: group.length,
-            totalArea: group.reduce((sum, area) => sum + (area.size * area.size), 0)
-          }
-        }
-        
-        console.log('[MemoryPalace] Created painted door params with AI properties:', paintedDoorParams)
-        
-        // Notify parent about painted door creation
-        if (onPaintedObjectCreated) {
-          console.log('[MemoryPalace] Notifying parent about painted door creation')
-          onPaintedObjectCreated(paintedDoorParams)
-        }
-      } else {
-        // Create painted object with proper type structure
-        const paintedObjectParams = {
-          name: finalName,
-          type: 'object',
-          information: finalInformation,
-          position: groupData.position,
-          isPaintedObject: true,
-          paintData: {
-            // Store only essential data after object creation - remove detailed stroke data
-            dimensions: groupData.dimensions,
-            canvasPosition: groupData.canvasCenter,
-            color: 'rgba(255, 0, 0, 0.9)', // Red for objects
-            aiGenerated: !!aiObjectProperties,
-            // Note: areas with full stroke data removed to reduce serialization overhead
-            strokeCount: group.length,
-            totalArea: group.reduce((sum, area) => sum + (area.size * area.size), 0)
-          }
-        }
-        
-        console.log('[MemoryPalace] Created painted object params with AI properties:', paintedObjectParams)
-        
-        // Notify parent about painted object creation
-        if (onPaintedObjectCreated) {
-          console.log('[MemoryPalace] Notifying parent about painted object creation')
-          onPaintedObjectCreated(paintedObjectParams)
-        }
-      }
-    })
-    
-    // Clear painted groups after processing to free memory and reduce serialization overhead
-    setPaintedGroups(new Map())
-    console.log('[MemoryPalace] Painted groups cleared after processing - stroke data cleanup complete')
-  }
-
-  const groupContiguousPaintAreas = (paintedAreas) => {
-    console.log('[MemoryPalace] Grouping contiguous paint areas:', paintedAreas.length)
-    
-    const groups = []
-    const visited = new Set()
-    const proximityThreshold = 75 // Pixels - areas within this distance are considered contiguous
-    
-    paintedAreas.forEach((area, index) => {
-      if (visited.has(index)) return
-      
-      // Start new group with current area
-      const group = []
-      const queue = [index]
-      
-      while (queue.length > 0) {
-        const currentIndex = queue.shift()
-        if (visited.has(currentIndex)) continue
-        
-        visited.add(currentIndex)
-        const currentArea = paintedAreas[currentIndex]
-        group.push(currentArea)
-        
-        // Find nearby areas to add to this group
-        paintedAreas.forEach((otherArea, otherIndex) => {
-          if (visited.has(otherIndex)) return
-          
-          const distance = Math.sqrt(
-            Math.pow(currentArea.center.x - otherArea.center.x, 2) + 
-            Math.pow(currentArea.center.y - otherArea.center.y, 2)
-          )
-          
-          if (distance <= proximityThreshold) {
-            queue.push(otherIndex)
-          }
-        })
-      }
-      
-      if (group.length > 0) {
-        groups.push(group)
-        console.log(`[MemoryPalace] Created group ${groups.length} with ${group.length} areas`)
-      }
-    })
-    
-    return groups
-  }
-
-  const createObjectFromPaintGroup = (group, groupNumber, type = 'objects') => {
-    console.log('[MemoryPalace] Creating', type, 'from paint group:', group.length, 'areas')
-    
-    // Calculate center position of the group
-    const centerX = group.reduce((sum, area) => sum + area.center.x, 0) / group.length
-    const centerY = group.reduce((sum, area) => sum + area.center.y, 0) / group.length
-    
-    // For latitude-adjusted areas, use the already adjusted sizes
-    // Fall back to a consistent size if areas don't have proper size information
-    const getAreaSize = (area) => area.size || 25;
-    
-    // Calculate bounding box of painted areas to determine geometry size
-    const minX = Math.min(...group.map(area => area.center.x - getAreaSize(area)))
-    const maxX = Math.max(...group.map(area => area.center.x + getAreaSize(area)))
-    const minY = Math.min(...group.map(area => area.center.y - getAreaSize(area)))
-    const maxY = Math.max(...group.map(area => area.center.y + getAreaSize(area)))
-    
-    const paintedWidth = maxX - minX
-    const paintedHeight = maxY - minY
-    
-    // Convert canvas dimensions to world scale
-    // Canvas is 4096x2048, representing 360° x 180° view
-    const canvasToWorldScale = 1000 / 2048 // Approximate scale factor
-    const worldWidth = Math.max(paintedWidth * canvasToWorldScale, 50) // Minimum 50 units
-    const worldHeight = Math.max(paintedHeight * canvasToWorldScale, 50) // Minimum 50 units
-    
-    // Calculate average world position
-    const avgWorldPos = group.reduce((sum, area) => {
-      return {
-        x: sum.x + area.worldPosition.x,
-        y: sum.y + area.worldPosition.y,
-        z: sum.z + area.worldPosition.z
-      }
-    }, { x: 0, y: 0, z: 0 })
-    
-    avgWorldPos.x /= group.length
-    avgWorldPos.y /= group.length
-    avgWorldPos.z /= group.length
-    
-    // Create different names and information based on type
-    const isObject = type === 'objects'
-    const objectData = {
-      name: isObject ? `Painted Object ${groupNumber}` : `Painted Door ${groupNumber}`,
-      information: isObject 
-        ? `Created from ${group.length} paint stroke${group.length > 1 ? 's' : ''}. Double-click to edit this painted memory.`
-        : `Created from ${group.length} paint stroke${group.length > 1 ? 's' : ''}. This door needs a destination room. Double-click to configure.`,
-      position: avgWorldPos,
-      canvasCenter: { x: centerX, y: centerY },
-      dimensions: {
-        width: worldWidth,
-        height: worldHeight,
-        canvasWidth: paintedWidth,
-        canvasHeight: paintedHeight
-      }
-    }
-    
-    console.log('[MemoryPalace] Created object data with dimensions:', objectData)
-    return objectData
-  }
-
-  // Helper function to convert canvas Y-coordinate to latitude in radians
-  const canvasYToLatitude = (canvasY, canvasHeight) => {
-    // Normalize Y to range [0, 1] where 0 = top (North pole), 1 = bottom (South pole)
-    const normalizedY = canvasY / canvasHeight;
-    // Convert to latitude: π/2 at top, -π/2 at bottom
-    const latitude = (Math.PI / 2) - (normalizedY * Math.PI);
-    return latitude;
-  };
-
-  // Helper function to calculate brush size adjusted for latitude
-  const calculateLatitudeAdjustedBrushSize = (baseBrushSize, canvasY, canvasHeight) => {
-    const latitude = canvasYToLatitude(canvasY, canvasHeight);
-    const cosLat = Math.cos(latitude);
-    
-    // Prevent division by zero at poles and extreme scaling
-    const clampedCosLat = Math.max(0.1, Math.abs(cosLat));
-    
-    // Calculate width and height adjustments separately to account for equirectangular distortion
-    // Width needs to be adjusted more at poles (inversely proportional to cosine of latitude)
-    // Height remains relatively consistent across latitudes
-    return {
-      width: baseBrushSize / clampedCosLat, // Width stretches horizontally at poles
-      height: baseBrushSize, // Height remains constant
-      size: baseBrushSize * Math.sqrt(1/clampedCosLat), // Original area-based adjustment for backward compatibility
-    };
-  };
-
-  const paintOnSkybox = (event) => {
-    console.log('[MemoryPalace] DEBUG: paintOnSkybox called', {
-      paintModeEnabled,
-      hasContext: !!paintContextRef.current,
-      hasCamera: !!cameraRef.current,
-      hasSkyboxSphere: !!skyboxSphereRef.current,
-      creationModeActive,
-      aiObjectProperties
-    })
-    
-    if (!paintModeEnabledRef.current || !paintContextRef.current || !cameraRef.current) return
-    
-    // Calculate mouse position in normalized device coordinates
-    const mouse = new THREE.Vector2()
-    mouse.x = (event.clientX / window.innerWidth) * 2 - 1
-    mouse.y = -(event.clientY / window.innerHeight) * 2 + 1
-    
-    console.log('[MemoryPalace] DEBUG: Mouse position', { clientX: event.clientX, clientY: event.clientY, normalizedX: mouse.x, normalizedY: mouse.y })
-    
-    // Create raycaster
-    const raycaster = new THREE.Raycaster()
-    raycaster.setFromCamera(mouse, cameraRef.current)
-    
-    // Check intersection with skybox sphere
-    if (!skyboxSphereRef.current) return
-    
-    const intersects = raycaster.intersectObject(skyboxSphereRef.current)
-    console.log('[MemoryPalace] DEBUG: Raycast intersections with skybox:', intersects.length)
-    
-    if (intersects.length > 0) {
-      const intersectionPoint = intersects[0].point
-      const uv = intersects[0].uv
-      
-      if (uv) {
-        // Convert UV coordinates to canvas coordinates
-        const canvas = paintCanvasRef.current
-        const context = paintContextRef.current
-        
-        // Convert UV coordinates to canvas coordinates
-        // UV coordinates: u=0 is left edge, v=0 is bottom edge
-        // Canvas coordinates: x=0 is left edge, y=0 is top edge
-        // Account for the 0.5 texture offset applied to paint texture (same as skybox textures)
-        let canvasX = ((uv.x + 0.5) % 1.0) * canvas.width  // Apply 180° offset compensation
-        let canvasY = (1.0 - uv.y) * canvas.height  // Flip Y-axis
-        
-        // Calculate latitude for this position
-        const latitude = canvasYToLatitude(canvasY, canvas.height);
-        
-        // Base brush size
-        const baseBrushSize = 25; // Reduced from 50 to 25 for smaller brush
-        
-        // Apply latitude adjustment to brush size to account for equirectangular distortion
-        const adjustedBrush = calculateLatitudeAdjustedBrushSize(baseBrushSize, canvasY, canvas.height);
-        
-        // Determine paint color based on AI properties or current paint mode type
-        let paintColor
-        let effectiveType = paintModeType
-        
-        // If in creation mode and AI has determined a type, use that
-        if (creationModeActive && aiObjectProperties?.type) {
-          effectiveType = aiObjectProperties.type === 'door' ? 'doors' : 'objects'
-          console.log('[MemoryPalace] AI determined type:', aiObjectProperties.type, 'effective paint type:', effectiveType)
-          
-          // Notify parent about paint type change if it differs from current
-          if (effectiveType !== paintModeType && onPaintTypeChange) {
-            onPaintTypeChange(effectiveType)
-          }
-        }
-        
-        // Set paint color based on effective type
-        paintColor = effectiveType === 'doors' 
-          ? 'rgba(0, 0, 255, 0.9)'  // Blue for doors
-          : 'rgba(255, 0, 0, 0.9)'  // Red for objects
-        
-        // Add visual feedback for AI-driven changes
-        if (creationModeActive && aiObjectProperties?.type) {
-          // Add slight glow effect for AI-determined paint
-          context.shadowColor = paintColor
-          context.shadowBlur = 10
-        } else {
-          context.shadowBlur = 0
-        }
-        
-        context.fillStyle = paintColor
-        context.beginPath()
-        
-        // Use elliptical brush to correctly account for equirectangular projection distortion
-        context.save();
-        context.translate(canvasX, canvasY);
-        
-        // Use proper elliptical drawing with correct aspect ratio for the latitude
-        // This creates wider strokes near poles and more circular strokes at the equator
-        context.scale(adjustedBrush.width / adjustedBrush.height, 1);
-        context.arc(0, 0, adjustedBrush.height/2, 0, Math.PI * 2);
-        context.restore();
-        
-        context.fill()
-        
-        // Reset shadow for next paint stroke
-        context.shadowBlur = 0
-        
-        // Store painted area info for later grouping (simplified)
-        const paintedArea = {
-          id: Date.now() + Math.random(),
-          center: { x: canvasX, y: canvasY },
-          size: adjustedBrush.size,
-          width: adjustedBrush.width,
-          height: adjustedBrush.height,
-          baseSize: baseBrushSize,
-          latitude: latitude,
-          distortionFactor: 1 / Math.max(0.1, Math.abs(Math.cos(latitude))),
-          color: paintColor,
-          worldPosition: intersectionPoint.clone(),
-          metadata: { 
-            name: aiObjectProperties?.name || `Painted Object ${paintedGroups.size + 1}`, 
-            info: aiObjectProperties?.information || aiObjectProperties?.description || 'Created with paint tool',
-            aiGenerated: !!aiObjectProperties
-          },
-          effectiveType: effectiveType
-        }
-        
-        setPaintedGroups(prevGroups => {
-          const newGroups = new Map(prevGroups)
-          newGroups.set(paintedArea.id, paintedArea)
-          return newGroups
-        })
-        
-        // Update texture
-        if (paintTextureRef.current) {
-          paintTextureRef.current.needsUpdate = true
-        }
-        exposePaintedAreasToLLM()
-        console.log('[MemoryPalace] Painted area created with AI integration:', paintedArea)
-      }
-    }
-  }
-
-  const getPaintedObjectAt = (event) => {
-    if (!paintModeEnabledRef.current) return null
-    
-    // Calculate mouse position and find painted object
-    const mouse = new THREE.Vector2()
-    mouse.x = (event.clientX / window.innerWidth) * 2 - 1
-    mouse.y = -(event.clientY / window.innerHeight) * 2 + 1
-    
-    const raycaster = new THREE.Raycaster()
-    raycaster.setFromCamera(mouse, cameraRef.current)
-    
-    if (!skyboxSphereRef.current) return null
-    
-    const intersects = raycaster.intersectObject(skyboxSphereRef.current)
-    
-    if (intersects.length > 0) {
-      const uv = intersects[0].uv
-      if (uv) {
-        const canvas = paintCanvasRef.current
-        let canvasX = ((uv.x + 0.5) % 1.0) * canvas.width  // Apply 180° offset compensation, same as paintOnSkybox
-        let canvasY = (1.0 - uv.y) * canvas.height
-        
-        // Calculate latitude at this position for brush size comparison
-        const latitude = canvasYToLatitude(canvasY, canvas.height)
-        
-        // Find closest painted area with elliptical hit detection
-        let closestArea = null
-        let minDistance = Infinity
-        
-        paintedGroups.forEach((area) => {
-          // Get horizontal and vertical distances separately
-          const dx = area.center.x - canvasX
-          const dy = area.center.y - canvasY
-          
-          // Use the stored width/height values for elliptical distance calculation
-          // If the area was created before our update, fall back to the size
-          const areaWidth = area.width || area.size || 25
-          const areaHeight = area.height || area.size || 25
-          
-          // Calculate normalized elliptical distance
-          // If we're within the ellipse defined by width and height, this will be < 1.0
-          const ellipticalDistance = Math.sqrt(
-            Math.pow(dx / areaWidth, 2) + 
-            Math.pow(dy / areaHeight, 2)
-          )
-          
-          // A point is inside the ellipse if the elliptical distance is < 0.5
-          // (since width/height represent the full diameter, not radius)
-          if (ellipticalDistance < 0.5 && ellipticalDistance < minDistance) {
-            minDistance = ellipticalDistance
-            closestArea = area
-          }
-        })
-        
-        return closestArea
-      }
-    }
-    
-    return null
-  }
-
-  // Skybox update function for room navigation
-  const updateSkyboxForRoom = (room) => {
-    if (!skyboxMaterialRef.current) {
-      console.warn('[MemoryPalace] Skybox material not ready for room update')
-      return
-    }
-
-    console.log('[MemoryPalace] Updating skybox for room:', room?.name || 'default')
-
-    // Skip texture loading if wireframe is enabled
-    // if (wireframeEnabled) {
-    //   console.log('[MemoryPalace] Wireframe enabled, skipping room texture loading')
-    //   return
-    // }
-
-    if (!wireframeEnabled && room?.imageUrl) {
-      // Load room-specific texture
-      console.log('[MemoryPalace] Loading room texture:', room.imageUrl)
-      const textureLoader = new THREE.TextureLoader()
-      
-      textureLoader.load(
-        room.imageUrl,
-        (texture) => {
-          console.log('[MemoryPalace] Room texture loaded successfully')
-          texture.mapping = THREE.EquirectangularReflectionMapping
-          texture.wrapS = THREE.RepeatWrapping
-          texture.wrapT = THREE.ClampToEdgeWrapping
-          texture.offset.x = 0.5 // Shift texture by 180 degrees
-          
-          // Dispose previous texture to prevent memory leaks
-          if (skyboxMaterialRef.current.map && skyboxMaterialRef.current.map !== texture) {
-            skyboxMaterialRef.current.map.dispose()
-          }
-          
-          skyboxMaterialRef.current.map = texture
-          skyboxMaterialRef.current.color.setHex(0xffffff)
-          skyboxMaterialRef.current.needsUpdate = true
-          
-          // Hide wireframe when texture loads
-          if (wireframeSphereRef.current && !wireframeEnabled) {
-            wireframeSphereRef.current.visible = false
-          }
-        },
-        undefined,
-        (error) => {
-          console.warn('[MemoryPalace] Failed to load room texture, keeping current:', error)
-        }
-      )
-    } else {
-      // No room-specific image, use default texture logic
-      console.log('[MemoryPalace] No room image, using default skybox')
-      // The default texture loading logic from initialization will be used
-    }
-  }
-
-  // Object rendering functions
-  const createObjectMarker = (obj) => {
-    if (!sceneRef.current) return null
-    console.log(`[MemoryPalace] 🏗️ SCENE: createObjectMarker called`, {
-      objectId: obj.id,
-      objectName: obj.name,
-      position: obj.position,
-      timestamp: new Date().toISOString()
-    })
-    
-    // Determine marker type based on explicit type property or fallback to targetRoomId
-    const isDoor = obj.type === 'door' || obj.targetRoomId !== undefined
-    const isPainted = obj.isPaintedObject || obj.isPaintedDoor
-    
-    console.log(`[MemoryPalace] 🎯 SCENE: marker type determined`, {
-      objectId: obj.id,
-      explicitType: obj.type,
-      isDoor,
-      isPainted,
-      markerType: isDoor ? 'door' : 'object',
-      hasPaintData: !!obj.paintData
-    })
-    
-    // Create appropriate geometry for marker type
-    let markerGeometry, markerMaterial
-    
-    if (isDoor) {
-      // For doors, check if painted and use dimensions from paint data
-      if (isPainted && obj.paintData?.dimensions) {
-        const width = Math.max(obj.paintData.dimensions.width, 100) // Minimum door width
-        const height = Math.max(obj.paintData.dimensions.height, 150) // Minimum door height
-        markerGeometry = new THREE.PlaneGeometry(width, height)
-        console.log(`[MemoryPalace] Using painted door geometry: ${width}x${height}`)
-      } else {
-        // Default door geometry
-        markerGeometry = new THREE.BoxGeometry(200, 300, 10)
-      }
-      markerMaterial = new THREE.MeshBasicMaterial({
-        color: 0xffd700,
-        transparent: true,
-        opacity: wireframeEnabled ? 1.0 : 0.0, // Invisible hit area
-        depthTest: false,
-        depthWrite: false,
-        wireframe: true,
-      })
-    } else {
-      // For objects, check if painted and use dimensions from paint data
-      if (obj.paintData?.dimensions) {
-        const width = Math.max(obj.paintData.dimensions.width, 50)
-        const height = Math.max(obj.paintData.dimensions.height, 50)
-        markerGeometry = new THREE.PlaneGeometry(width, height)
-        console.log(`[MemoryPalace] Using painted object geometry: ${width}x${height}`)
-      } else {
-        // Default object geometry
-        markerGeometry = new THREE.SphereGeometry(50, 8, 6)
-      }
-      markerMaterial = new THREE.MeshBasicMaterial({
-        color: 0x4dabf7,
-        transparent: true,
-        opacity: wireframeEnabled ? 1.0 : 0.0, // Invisible hit area
-        depthTest: false,
-        depthWrite: false,
-        wireframe: true,
-      })
-    }
-    
-    const marker = new THREE.Mesh(markerGeometry, markerMaterial)
-    
-    // Position marker just outside sphere surface (radius 500)
-    const baseRadius = 500
-    
-    if (obj.position?.x !== undefined && obj.position?.y !== undefined && obj.position?.z !== undefined) {
-      const direction = new THREE.Vector3(obj.position.x, obj.position.y, obj.position.z).normalize()
-      marker.position.copy(direction.multiplyScalar(baseRadius))
-      
-      // Orient plane geometries to face toward camera (center)
-      if (isPainted && obj.paintData?.dimensions) {
-        // For plane geometries, ensure they face toward the camera
-        const lookDirection = new THREE.Vector3(0, 0, 0).sub(marker.position).normalize()
-        marker.lookAt(marker.position.clone().add(lookDirection))
-      } else if (isDoor) {
-        // For doors (including non-painted), orient to face the user's view
-        const lookDirection = new THREE.Vector3(0, 0, 0).sub(marker.position).normalize()
-        marker.lookAt(marker.position.clone().add(lookDirection))
-      }
-      // Objects with sphere geometry don't need special orientation
-      
-      console.log(`[MemoryPalace] 📍 SCENE: marker positioned using object coordinates`, {
-        objectId: obj.id,
-        objectPosition: obj.position,
-        markerPosition: marker.position
-      })
-    } else {
-      // Fallback to default position (facing forward)
-      console.warn(`[MemoryPalace] ⚠️ SCENE: object has no 3D position, using default`, {
-        objectId: obj.id,
-        objectName: obj.name
-      })
-      marker.position.set(baseRadius, 0, 0) // Forward facing
-    }
-    
-    // Store object data in userData
-    marker.userData = {
-      objectData: obj,
-      isDoor: isDoor,
-      objectId: obj.id,
-      originalScale: marker.scale.clone()
-    }
-    
-    // Create particle system for this marker
-    if (particleManagerRef.current) {
-      const width = Math.max(obj.paintData?.dimensions?.width, 50)
-      const height = Math.max(obj.paintData?.dimensions?.height, 50)
-      console.log(`[MemoryPalace] ✨ SCENE: creating particle system for marker`, {
-        objectId: obj.id,
-        markerPosition: marker.position,
-        width,
-        height,
-      })
-      
-      
-      
-      const particleSystem = particleManagerRef.current.createParticleSystem(marker.position, isDoor, obj.id, {
-        width,
-        height,
-        shape: "box",
-        density: 0.5,
-        swirlSpeed: 1.7,
-        swirlTightness: 0.0,
-        twinkleSpeed: 0,
-        twinkleIntensity: 2.55
-      })
-      sceneRef.current.add(particleSystem)
-      
-      // Store particle system reference
-      marker.userData.particleSystem = particleSystem
-    }
-    
-    sceneRef.current.add(marker)
-    
-    console.log(`[MemoryPalace] ✅ SCENE: object marker added to scene`, {
-      objectId: obj.id,
-      objectName: obj.name,
-      markerPosition: marker.position,
-      isDoor,
-      hasParticleSystem: !!marker.userData.particleSystem,
-      timestamp: new Date().toISOString()
-    })
-    
-    return marker
-  }
-
-  const updateObjectMarkers = (newObjects) => {
-    if (!sceneRef.current) return
-    
-    console.log(`[MemoryPalace] 🎬 SCENE: updateObjectMarkers called`, {
-      newObjectsCount: newObjects.length,
-      newObjectIds: newObjects.map(obj => obj.id),
-      newObjectNames: newObjects.map(obj => obj.name),
-      timestamp: new Date().toISOString()
-    })
-    
-    // Get current markers
-    const currentMarkers = objectMarkersRef.current
-    const currentObjectIds = new Set(Array.from(currentMarkers.keys()))
-    const newObjectIds = new Set(newObjects.map(obj => obj.id))
-    
-    console.log(`[MemoryPalace] 🔍 SCENE: marker comparison`, {
-      currentMarkerCount: currentMarkers.size,
-      currentObjectIds: Array.from(currentObjectIds),
-      newObjectIds: Array.from(newObjectIds),
-      markersToRemove: Array.from(currentObjectIds).filter(id => !newObjectIds.has(id)),
-      markersToAdd: Array.from(newObjectIds).filter(id => !currentObjectIds.has(id))
-    })
-    
-    // Remove markers for objects that no longer exist
-    currentObjectIds.forEach(objectId => {
-      if (!newObjectIds.has(objectId)) {
-        console.log(`[MemoryPalace] 🗑️ SCENE: removing marker for deleted object`, { objectId })
-        const marker = currentMarkers.get(objectId)
-        if (marker) {
-          // Remove particle system
-          if (marker.userData.particleSystem && particleManagerRef.current) {
-            sceneRef.current.remove(marker.userData.particleSystem)
-            particleManagerRef.current.removeParticleSystem(objectId)
-          }
-          
-          // Remove marker from scene
-          sceneRef.current.remove(marker)
-          
-          // Dispose resources
-          if (marker.geometry) marker.geometry.dispose()
-          if (marker.material) marker.material.dispose()
-          
-          currentMarkers.delete(objectId)
-          console.log(`[MemoryPalace] ✅ SCENE: marker removed from scene`, { objectId })
-        }
-      }
-    })
-    
-    // Add markers for new objects
-    newObjects.forEach(obj => {
-      if (!currentMarkers.has(obj.id)) {
-        console.log(`[MemoryPalace] ➕ SCENE: adding marker for new object`, {
-          objectId: obj.id,
-          objectName: obj.name,
-          position: obj.position
-        })
-        const marker = createObjectMarker(obj)
-        if (marker) {
-          currentMarkers.set(obj.id, marker)
-          console.log(`[MemoryPalace] ✅ SCENE: new marker added to scene`, {
-            objectId: obj.id,
-            objectName: obj.name,
-            markerPosition: marker.position
-          })
-        }
-      } else {
-        // Update existing marker if needed (position, properties)
-        const marker = currentMarkers.get(obj.id)
-        if (marker && marker.userData.objectData) {
-          // Check if position changed
-          const oldPos = marker.userData.objectData.position
-          const newPos = obj.position
-          if (oldPos && newPos && 
-              (oldPos.x !== newPos.x || oldPos.y !== newPos.y || oldPos.z !== newPos.z)) {
-            console.log(`[MemoryPalace] 🔄 SCENE: updating position for existing object`, {
-              objectId: obj.id,
-              objectName: obj.name,
-              oldPosition: oldPos,
-              newPosition: newPos
-            })
-            
-            // Update marker position
-            const isDoor = obj.targetRoomId !== undefined
-            if (isDoor) {
-              // For doors, use exact position coordinates
-              marker.position.set(newPos.x, newPos.y, newPos.z)
-              
-              // Orient door to face the user's view (toward center)
-              const lookDirection = new THREE.Vector3(0, 0, 0).sub(marker.position).normalize()
-              marker.lookAt(marker.position.clone().add(lookDirection))
-            } else {
-              // For objects, extend to sphere surface
-              const baseRadius = 500
-              const direction = new THREE.Vector3(newPos.x, newPos.y, newPos.z).normalize()
-              marker.position.copy(direction.multiplyScalar(baseRadius))
-            }
-            
-            // Update particle system position
-            if (marker.userData.particleSystem) {
-              marker.userData.particleSystem.position.copy(marker.position)
-            }
-            
-            console.log(`[MemoryPalace] ✅ SCENE: object position updated in scene`, {
-              objectId: obj.id,
-              newMarkerPosition: marker.position
-            })
-          }
-          
-          // Update userData with new object data
-          marker.userData.objectData = obj
-        }
-      }
-    })
-    
-    console.log(`[MemoryPalace] ✅ SCENE: updateObjectMarkers completed`, {
-      totalActiveMarkers: currentMarkers.size,
-      activeMarkerIds: Array.from(currentMarkers.keys()),
-      timestamp: new Date().toISOString()
-    })
-  }
-
-  const startObjectAnimation = () => {
-    if (animationFrameRef.current) return // Already animating
-    
-    const animate = () => {
-      // Update particle systems
-      if (particleManagerRef.current) {
-        particleManagerRef.current.updateParticleSystems()
-      }
-      
-      // Animate object markers (subtle pulsing)
-      // objectMarkersRef.current.forEach(marker => {
-      //   if (marker.userData.originalScale) {
-      //     const time = Date.now() * 0.001
-      //     const scale = 1.0 + Math.sin(time * 2) * 0.1 // Gentle pulsing
-      //     marker.scale.copy(marker.userData.originalScale).multiplyScalar(scale)
-      //   }
-      // })
-      
-      // Animate off-screen indicators
-      animateObjectIndicators()
-      
-      animationFrameRef.current = requestAnimationFrame(animate)
-    }
-    
-    animate()
-  }
-
-  const stopObjectAnimation = () => {
-    if (animationFrameRef.current) {
-      cancelAnimationFrame(animationFrameRef.current)
-      animationFrameRef.current = null
-    }
-  }
-
-  // Off-screen indicator functions
-  const createObjectIndicator = (obj, direction, isInView) => {
-    if (!sceneRef.current || !cameraRef.current) return null
-
-    console.log(`[MemoryPalace] 🎯 Creating object indicator:`, {
-      objectId: obj.id,
-      objectName: obj.name,
-      isInView: isInView,
-      direction: direction
-    })
-
-    // Create arrow geometry pointing in the direction of the object
-    const arrowGeometry = new THREE.ConeGeometry(0.08, 0.4, 4)
-    // Align cone's forward axis (+Z) with look direction
-    arrowGeometry.rotateX(Math.PI / 2)
-    const isDoor = obj.targetRoomId !== undefined
-    
-    // Use same color scheme as object markers
-    const arrowMaterial = new THREE.MeshBasicMaterial({
-      color: 0xffffff,
-      transparent: true,
-      opacity: 0.0,
-      depthTest: false,
-      depthWrite: false
-    })
-
-    const arrow = new THREE.Mesh(arrowGeometry, arrowMaterial)
-    
-    const camera = cameraRef.current
-    const fixedDistance = 8
-    
-    // Calculate direction from camera to object in camera space
-    const objectWorldPos = new THREE.Vector3(obj.position.x, obj.position.y, obj.position.z)
-    const directionFromCamera = objectWorldPos.clone().sub(camera.position).normalize()
-    
-    let indicatorPosition
-    
-    if (isInView) {
-      // Object is in view - position indicator at object center
-      const tempVector = directionFromCamera.clone()
-      tempVector.project(camera)
-      
-      // Keep original screen position (don't clamp to edges)
-      tempVector.z = 0.1
-      
-      // Convert back to world space at fixed distance
-      tempVector.unproject(camera)
-      const indicatorDirection = tempVector.sub(camera.position).normalize()
-      indicatorPosition = camera.position.clone().add(indicatorDirection.multiplyScalar(fixedDistance))
-    } else {
-      // Object is out of view - clamp to screen edges
-      const tempVector = directionFromCamera.clone()
-      tempVector.project(camera)
-      
-      // Find closest edge position
-      const margin = 0.85
-      const absX = Math.abs(tempVector.x)
-      const absY = Math.abs(tempVector.y)
-      
-      // Determine which edge is closest and position on that edge
-      if (absX > absY) {
-        // Object is more to the left/right - position on vertical edges
-        tempVector.x = tempVector.x > 0 ? margin : -margin
-        tempVector.y = Math.max(-margin, Math.min(margin, tempVector.y))
-      } else {
-        // Object is more up/down - position on horizontal edges  
-        tempVector.y = tempVector.y > 0 ? margin : -margin
-        tempVector.x = Math.max(-margin, Math.min(margin, tempVector.x))
-      }
-      tempVector.z = 0.1
-      
-      // Convert back to world space at fixed distance
-      tempVector.unproject(camera)
-      const indicatorDirection = tempVector.sub(camera.position).normalize()
-      indicatorPosition = camera.position.clone().add(indicatorDirection.multiplyScalar(fixedDistance))
-    }
-    
-    arrow.position.copy(indicatorPosition)
-
-    // Point arrow toward the object from its current position
-    const forward = new THREE.Vector3(0, 0, 1)
-    let directionToObject = objectWorldPos.clone().sub(indicatorPosition).normalize()
-    const cameraForward = new THREE.Vector3()
-    camera.getWorldDirection(cameraForward)
-    if (directionFromCamera.dot(cameraForward) < 0) {
-      // Object is behind camera; use projected indicator direction so arrow
-      // points from screen edge toward the object
-      const tempDir = directionFromCamera.clone()
-      tempDir.project(camera)
-      if (!isInView) {
-        // Only adjust direction for out-of-view objects
-        directionToObject = camera.position.clone().sub(indicatorPosition).normalize().negate()
-      }
-    }
-    arrow.quaternion.setFromUnitVectors(forward, directionToObject)
-    
-    // Add distance text sprite
-    const distance = camera.position.distanceTo(objectWorldPos)
-    const distanceText = Math.round(distance).toString() + 'm'
-    
-    const canvas = document.createElement('canvas')
-    const context = canvas.getContext('2d')
-    canvas.width = 128
-    canvas.height = 64
-    
-    context.fillStyle = 'white'
-    context.font = 'bold 10px Arial'
-    context.textAlign = 'center'
-    context.fillText(obj.name, 64, 20)
-    
-    const texture = new THREE.CanvasTexture(canvas)
-    const spriteMaterial = new THREE.SpriteMaterial({ 
-      map: texture,
-      transparent: true,
-      opacity: 0.0,
-      depthTest: false
-    })
-    const sprite = new THREE.Sprite(spriteMaterial)
-    sprite.scale.set(0.75, 0.375, 1)
-    
-    // Position sprite slightly behind arrow
-    const spriteDirection = camera.position.clone().sub(indicatorPosition).normalize()
-    const spriteOffset = spriteDirection.multiplyScalar(-0.5)
-    sprite.position.copy(indicatorPosition.clone().add(spriteOffset))
-    
-    // Store references
-    arrow.userData = {
-      objectData: obj,
-      objectId: obj.id,
-      isDoor: isDoor,
-      sprite: sprite,
-      originalOpacity: 1.0,
-      originalPosition: indicatorPosition.clone(),
-      originalSpritePosition: sprite.position.clone()
-    }
-    
-    sprite.userData = {
-      objectId: obj.id,
-      arrow: arrow
-    }
-    
-    sceneRef.current.add(arrow)
-    sceneRef.current.add(sprite)
-    
-    return { arrow, sprite }
-  }
-
-  const updateObjectIndicators = (objects) => {
-    if (!sceneRef.current || !cameraRef.current) return
-
-    const camera = cameraRef.current
-    const frustum = new THREE.Frustum()
-    const cameraMatrix = new THREE.Matrix4()
-    cameraMatrix.multiplyMatrices(camera.projectionMatrix, camera.matrixWorldInverse)
-    frustum.setFromProjectionMatrix(cameraMatrix)
-
-    // Clear existing indicators
-    const currentIndicators = offScreenIndicatorsRef.current
-    currentIndicators.forEach((indicator, objectId) => {
-      if (indicator.arrow) {
-        sceneRef.current.remove(indicator.arrow)
-        if (indicator.arrow.geometry) indicator.arrow.geometry.dispose()
-        if (indicator.arrow.material) indicator.arrow.material.dispose()
-      }
-      if (indicator.sprite) {
-        sceneRef.current.remove(indicator.sprite)
-        if (indicator.sprite.material && indicator.sprite.material.map) {
-          indicator.sprite.material.map.dispose()
-        }
-        if (indicator.sprite.material) indicator.sprite.material.dispose()
-      }
-    })
-    currentIndicators.clear()
-
-    // Create indicators for ALL objects, positioning based on visibility
-    objects.forEach(obj => {
-      if (!obj.position) return
-
-      const objectPosition = new THREE.Vector3(obj.position.x, obj.position.y, obj.position.z)
-      const isInView = frustum.containsPoint(objectPosition)
-      
-      // Calculate direction from camera to object
-      const direction = objectPosition.clone().sub(camera.position)
-      
-      // Create indicator (positioned at center if in view, at edge if not)
-      const indicator = createObjectIndicator(obj, direction, isInView)
-      if (indicator) {
-        currentIndicators.set(obj.id, indicator)
-      }
-    })
-
-    console.log(`[MemoryPalace] 🎯 Updated object indicators:`, {
-      totalObjects: objects.length,
-      indicatorsCreated: currentIndicators.size,
-      timestamp: new Date().toISOString(),
-      objects,
-      currentIndicators,
-    })
-  }
-
-  const animateObjectIndicators = () => {
-    const currentIndicators = offScreenIndicatorsRef.current
-    const camera = cameraRef.current
-
-    if (!camera) return
-
-    currentIndicators.forEach((indicator, objectId) => {
-      if (indicator.arrow && indicator.sprite) {
-        // Update positions to maintain proper placement as camera moves
-        const obj = indicator.arrow.userData.objectData
-        const objectWorldPos = new THREE.Vector3(obj.position.x, obj.position.y, obj.position.z)
-        const directionFromCamera = objectWorldPos.clone().sub(camera.position).normalize()
-        
-        // Check if object is in view to determine positioning
-        const frustum = new THREE.Frustum()
-        const cameraMatrix = new THREE.Matrix4()
-        cameraMatrix.multiplyMatrices(camera.projectionMatrix, camera.matrixWorldInverse)
-        frustum.setFromProjectionMatrix(cameraMatrix)
-        const isInView = frustum.containsPoint(objectWorldPos)
-        
-        const tempVector = directionFromCamera.clone()
-        tempVector.project(camera)
-        
-        let indicatorPosition
-        
-        if (isInView) {
-          // Object is in view - position at object center
-          tempVector.z = 0.1
-          tempVector.unproject(camera)
-          const indicatorDirection = tempVector.sub(camera.position).normalize()
-          const fixedDistance = 8
-          indicatorPosition = camera.position.clone().add(indicatorDirection.multiplyScalar(fixedDistance))
-        } else {
-          // Object is out of view - clamp to screen edges
-          const margin = 0.85
-          const absX = Math.abs(tempVector.x)
-          const absY = Math.abs(tempVector.y)
-          
-          if (absX > absY) {
-            tempVector.x = tempVector.x > 0 ? margin : -margin
-            tempVector.y = Math.max(-margin, Math.min(margin, tempVector.y))
-          } else {
-            tempVector.y = tempVector.y > 0 ? margin : -margin
-            tempVector.x = Math.max(-margin, Math.min(margin, tempVector.x))
-          }
-          tempVector.z = 0.1
-          
-          tempVector.unproject(camera)
-          const indicatorDirection = tempVector.sub(camera.position).normalize()
-          const fixedDistance = 8
-          indicatorPosition = camera.position.clone().add(indicatorDirection.multiplyScalar(fixedDistance))
-        }
-        
-        // Update arrow position and rotation
-        indicator.arrow.position.copy(indicatorPosition)
-        const forward = new THREE.Vector3(0, 0, 1)
-        let directionToObject = objectWorldPos.clone().sub(indicatorPosition).normalize()
-        const cameraForward = new THREE.Vector3()
-        camera.getWorldDirection(cameraForward)
-        if (directionFromCamera.dot(cameraForward) < 0) {
-          directionToObject = tempVector.sub(camera.position).normalize().negate()
-        }
-        indicator.arrow.quaternion.setFromUnitVectors(forward, directionToObject)
-        
-        // Update sprite position
-        const spriteOffset = tempVector.sub(camera.position).normalize().multiplyScalar(-0.5)
-        indicator.sprite.position.copy(indicatorPosition.clone().add(spriteOffset))
-        
-        // Simple static opacity (no pulsing/floating animations)
-        indicator.arrow.material.opacity = indicator.arrow.userData.originalOpacity || 0.8
-        indicator.sprite.material.opacity = 0.9
-      }
-    })
-  }
-
-  // Nipple.js functions
-  const initializeNipple = () => {
-    if (nippleManagerRef.current || !nippleEnabled) return
-
-    console.log('Initializing nipple.js controls...')
-    
-    // Create nipple container if it doesn't exist
-    if (!nippleContainerRef.current) {
-      const container = document.createElement('div')
-      container.className = 'nipple-container'
-      document.body.appendChild(container)
-      nippleContainerRef.current = container
-    }
-
-    // Initialize nipple manager
-    const manager = nipplejs.create({
-      zone: nippleContainerRef.current,
-      mode: 'static',
-      position: { left: '50%', top: '50%' },
-      color: 'rgba(255, 255, 255, 0.8)',
-      size: 100
-    })
-
-    nippleManagerRef.current = manager
-
-    // Handle nipple events
-    manager.on('move', (evt, data) => {
-      if (!cameraRef.current) return
-
-      const force = data.force
-      const angle = data.angle.radian
-      
-      // Convert nipple input to camera rotation
-      const rotationSpeed = 1000 * force // Increased for better responsiveness
-      const deltaX = Math.cos(angle) * rotationSpeed
-      const deltaY = Math.sin(angle) * rotationSpeed
-      
-      // Update yaw and pitch using prototype's approach
-      cameraRotationRef.current.yaw -= deltaX   // Horizontal rotation (left/right) - fixed direction
-      cameraRotationRef.current.pitch += deltaY // Vertical rotation (up/down) - fixed direction
-      
-      // Constrain horizontal rotation to 90% freedom (prevent viewing directly behind) - from prototype
-      cameraRotationRef.current.yaw = Math.max(-135, Math.min(135, cameraRotationRef.current.yaw))
-      
-      // Clamp vertical rotation - from prototype
-      cameraRotationRef.current.pitch = Math.max(-85, Math.min(85, cameraRotationRef.current.pitch))
-      
-      // Apply rotation using spherical coordinate approach (adjusted for forward-facing)
-      const phi = THREE.MathUtils.degToRad(90 - cameraRotationRef.current.pitch)
-      const theta = THREE.MathUtils.degToRad(cameraRotationRef.current.yaw)
-      
-      // Rotate coordinate system so 0,0 faces forward (positive X)
-      const x = 500 * Math.sin(phi) * Math.cos(theta)
-      const y = 500 * Math.cos(phi)
-      const z = 500 * Math.sin(phi) * Math.sin(theta)
-      
-      cameraRef.current.lookAt(x, y, z)
-    })
-
-    console.log('Nipple.js controls initialized')
-  }
-
-  const cleanupNipple = () => {
-    if (nippleManagerRef.current) {
-      nippleManagerRef.current.destroy()
-      nippleManagerRef.current = null
-    }
-    
-    if (nippleContainerRef.current) {
-      document.body.removeChild(nippleContainerRef.current)
-      nippleContainerRef.current = null
-    }
-  }
+  // Hooks for modularized logic
+  const { cameraRotationRef, applyRotation } = useCameraControls(cameraRef)
+  const {
+    paintCanvasRef,
+    paintTextureRef,
+    paintContextRef,
+    paintedGroups,
+    setPaintedGroups,
+    paintModeEnabledRef,
+    initializePaintCanvas,
+    enablePaintMode,
+    disablePaintMode,
+    cleanupStrokeData,
+    exposePaintedAreasToLLM,
+    processPaintedAreasIntoObjects,
+    paintOnSkybox,
+    getPaintedObjectAt
+  } = usePaintMode({
+    sceneRef,
+    cameraRef,
+    skyboxMaterialRef,
+    skyboxSphereRef,
+    creationModeActive,
+    paintModeType,
+    onPaintTypeChange,
+    onPaintedObjectCreated,
+    onPaintedAreasChange,
+    aiObjectProperties
+  })
+  const { initializeNipple, cleanupNipple } = useJoystick({
+    nippleEnabled,
+    cameraRef,
+    applyRotation
+  })
 
   // Update camera FOV dynamically
   const updateCameraFov = (newFov) => {
@@ -1675,15 +362,10 @@ const MemoryPalace = forwardRef(({
     const keyboardSensitivity = settingsManagerRef.current.get('keyboardSensitivity') || 0.02
     const dampingFactor = 0.85
     const maxVerticalAngle = Math.PI / 2 - 0.1 // Slight margin from straight up/down
-    
-    // Initialize camera rotation state from ref (facing forward in equirectangular space)
-    let yaw = cameraRotationRef.current.yaw || 0   // Start looking forward (center of equirectangular)
-    let pitch = cameraRotationRef.current.pitch || 0 // Start looking straight ahead
-    
+
     // Set initial camera position and look direction
     camera.position.set(0, 0, 0)
-    // Face forward (positive X direction) which is the center of equirectangular textures
-    camera.lookAt(1, 0, 0)
+    applyRotation(0, 0)
     
     // Keyboard state
     const keys = {
@@ -1718,39 +400,16 @@ const MemoryPalace = forwardRef(({
         
         const deltaX = (event.clientX - lastMouseX) * mouseSensitivity
         const deltaY = (event.clientY - lastMouseY) * mouseSensitivity
-        
-        // Update yaw and pitch using prototype's approach
-        yaw -= deltaX * 50   // Horizontal rotation (left/right) - fixed direction and scaling
-        pitch += deltaY * 50 // Vertical rotation (up/down) - fixed direction and scaling
-        
-        // Constrain horizontal rotation to 90% freedom (prevent viewing directly behind) - from prototype
-        yaw = Math.max(-135, Math.min(135, yaw))
-        
-        // Clamp vertical rotation - from prototype
-        pitch = Math.max(-85, Math.min(85, pitch))
-        
-        // Store back to ref
-        cameraRotationRef.current.yaw = yaw
-        cameraRotationRef.current.pitch = pitch
-        
-        // Apply rotation using spherical coordinate approach (adjusted for forward-facing)
-        const phi = THREE.MathUtils.degToRad(90 - pitch)
-        const theta = THREE.MathUtils.degToRad(yaw)
-        
-        // Rotate coordinate system so 0,0 faces forward (positive X)
-        const x = 500 * Math.sin(phi) * Math.cos(theta)
-        const y = 500 * Math.cos(phi)  
-        const z = 500 * Math.sin(phi) * Math.sin(theta)
-        
-        camera.lookAt(x, y, z)
-        
+
+        applyRotation(deltaX * 50, deltaY * 50)
+
         // Store velocity for momentum (optional future enhancement)
         rotationVelocityX = deltaY * 0.1
         rotationVelocityY = deltaX * 0.1
-        
+
         lastMouseX = event.clientX
         lastMouseY = event.clientY
-        
+
         // If we've moved significantly, it's not a click
         if (Math.abs(deltaX) > 0.01 || Math.abs(deltaY) > 0.01) {
           isClick = false
@@ -2111,35 +770,12 @@ const MemoryPalace = forwardRef(({
         
         const deltaX = (event.touches[0].clientX - lastMouseX) * touchSensitivity
         const deltaY = (event.touches[0].clientY - lastMouseY) * touchSensitivity
-        
-        // Update yaw and pitch using prototype's approach
-        yaw -= deltaX * 50   // Horizontal rotation (left/right) - fixed direction and scaling
-        pitch += deltaY * 50 // Vertical rotation (up/down) - fixed direction and scaling
-        
-        // Constrain horizontal rotation to 90% freedom (prevent viewing directly behind) - from prototype
-        yaw = Math.max(-135, Math.min(135, yaw))
-        
-        // Clamp vertical rotation - from prototype
-        pitch = Math.max(-85, Math.min(85, pitch))
-        
-        // Store back to ref
-        cameraRotationRef.current.yaw = yaw
-        cameraRotationRef.current.pitch = pitch
-        
-        // Apply rotation using spherical coordinate approach (adjusted for forward-facing)
-        const phi = THREE.MathUtils.degToRad(90 - pitch)
-        const theta = THREE.MathUtils.degToRad(yaw)
-        
-        // Rotate coordinate system so 0,0 faces forward (positive X)
-        const x = 500 * Math.sin(phi) * Math.cos(theta)
-        const y = 500 * Math.cos(phi)
-        const z = 500 * Math.sin(phi) * Math.sin(theta)
-        
-        camera.lookAt(x, y, z)
-        
+
+        applyRotation(deltaX * 50, deltaY * 50)
+
         lastMouseX = event.touches[0].clientX
         lastMouseY = event.touches[0].clientY
-        
+
         // If we've moved significantly, it's not a tap
         if (Math.abs(deltaX) > 0.01 || Math.abs(deltaY) > 0.01) {
           isClick = false
@@ -2244,29 +880,7 @@ const MemoryPalace = forwardRef(({
 
         // Apply rotation
         if (deltaX !== 0 || deltaY !== 0) {
-          yaw += deltaX
-          pitch += deltaY
-
-          // Constrain horizontal rotation to 90% freedom (prevent viewing directly behind) - from prototype
-          yaw = Math.max(-135, Math.min(135, yaw))
-          
-          // Clamp vertical rotation - from prototype
-          pitch = Math.max(-85, Math.min(85, pitch))
-          
-          // Store back to ref
-          cameraRotationRef.current.yaw = yaw
-          cameraRotationRef.current.pitch = pitch
-          
-          // Apply rotation using spherical coordinate approach (adjusted for forward-facing)
-          const phi = THREE.MathUtils.degToRad(90 - pitch)
-          const theta = THREE.MathUtils.degToRad(yaw)
-          
-          // Rotate coordinate system so 0,0 faces forward (positive X)
-          const x = 500 * Math.sin(phi) * Math.cos(theta)
-          const y = 500 * Math.cos(phi)
-          const z = 500 * Math.sin(phi) * Math.sin(theta)
-          
-          camera.lookAt(x, y, z)
+          applyRotation(deltaX, deltaY)
         }
       }
     }
@@ -2307,11 +921,6 @@ const MemoryPalace = forwardRef(({
 
     animate()
     console.log('Memory Palace scene initialized successfully')
-    
-    // Initialize nipple if enabled
-    if (nippleEnabled) {
-      setTimeout(initializeNipple, 100) // Small delay to ensure DOM is ready
-    }
 
     // Start object animation loop
     startObjectAnimation()
@@ -2479,15 +1088,6 @@ const MemoryPalace = forwardRef(({
       }
     }
   }, [wireframeEnabled])
-
-  // Handle nipple toggle
-  useEffect(() => {
-    if (nippleEnabled) {
-      initializeNipple()
-    } else {
-      cleanupNipple()
-    }
-  }, [nippleEnabled])
 
   // Handle paint mode toggle
   useEffect(() => {
