+++ src/hooks/index.js
@@ -2,5 +2,7 @@
  * Hooks Index - Export all custom hooks
  */
 
-export { useAnthropicStream } from './useAnthropicStream.js'
-export { default } from './useAnthropicStream.js'
+export { useAnthropicStream } from './useAnthropicStream.js';
+export { default as useCameraControls } from './useCameraControls.js';
+export { default as useJoystick } from './useJoystick.js';
+export { default as usePaintMode } from './usePaintMode.js';
